add hook dokterChat
